pyproject.toml
[project]
name = "flowdesk"
version = "0.1.0"
description = ""
authors = [
    {name = "MaxBakshaev",email = "risen.2011@mail.ru"}
]
readme = "README.md"
requires-python = "3.12.10"
dependencies = [
    "fastapi (>=0.123.9,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "aiogram (>=3.23.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
